fix highlighter type failure

+++ Browser/utils/time_conversion.py
@@ -2,4 +2,4 @@ from robot.utils import timestr_to_secs  # type: ignore
 
 
 def timestr_to_millisecs(time_str: str) -> int:
-    return timestr_to_secs(time_str) * 1000
+    return int(timestr_to_secs(time_str) * 1000)
